+++ sndata/base_classes.py
@@ -10,7 +10,6 @@ docs.
 
 import functools
 import shutil
-import warnings
 from typing import List
 from typing import Union